[BUGFIX] Do not attempt to delete missing or deleted file
Prevents calling methods that will fail for obvious reasons.

+++ Classes/Mapping/FalMapping.php
@@ -79,7 +79,7 @@ class FalMapping extends AbstractMapping
                     $this->performSanityCheckBeforeDeletion($record);
                 }
 
-                if ($object) {
+                if ($object && !$object->isMissing() && !$object->isDeleted()) {
                     $object->delete();
                     $repository->remove($object);
                 }
